# openapi.yaml
openapi: 3.0.3
info:
  title: SAP MCP Server API
  description: API for SAP Business Partner and Product searches
  version: 1.0.0
  contact:
    name: SAP MCP Server
    email: support@motiveminds.com

servers:
  - url: https://257f376919a3.ngrok-free.app
    description: Production server via ngrok

security:
  - ApiKeyAuth: []

paths:
  /health:
    get:
      summary: Health Check
      description: Check if the service is running
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  service:
                    type: string
                    example: "SAP MCP Server"
                  timestamp:
                    type: number
                  uptime:
                    type: number

  /api/sap/customer/search:
    post:
      summary: Search SAP Customers
      description: Search for customers by description in SAP Business Partner system
      operationId: searchCustomers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
              properties:
                description:
                  type: string
                  description: Customer description to search for
                  example: "ABC Corporation"
                search_field:
                  type: string
                  description: Field to search in
                  enum: ["CustomerName", "CustomerFullName"]
                  default: "CustomerName"
                exact_match:
                  type: boolean
                  description: Whether to search for exact match
                  default: false
                max_results:
                  type: integer
                  description: Maximum number of results
                  minimum: 1
                  maximum: 100
                  default: 10
                product:
                  type: string
                  description: Specific product ID for direct lookup
                key_mode:
                  type: string
                  description: Key access mode
                  enum: ["segment", "paren", "auto"]
                  default: "auto"
                service_version:
                  type: string
                  description: Service version
                  default: "0002"
                base_url:
                  type: string
                  description: Optional custom base URL
      responses:
        '200':
          description: Successful search
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      mode:
                        type: string
                      search_query:
                        type: string
                      language:
                        type: string
                      found_products:
                        type: integer
                      products:
                        type: array
                        items:
                          type: object
                          properties:
                            Product:
                              type: string
                              description: Product ID
                            ProductDescription:
                              type: string
                              description: Product description
                            Language:
                              type: string
                              description: Language code
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/sap/generic:
    post:
      summary: Generic SAP API Call
      description: Make generic calls to SAP OData services
      operationId: genericSapCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - service_path
              properties:
                service_path:
                  type: string
                  description: SAP OData service path
                  example: "sap/opu/odata/sap/API_SALES_ORDER_SRV/A_SalesOrder"
                method:
                  type: string
                  description: HTTP method
                  enum: ["GET", "POST", "PUT", "DELETE"]
                  default: "GET"
                headers:
                  type: object
                  description: Additional headers
                  additionalProperties:
                    type: string
                params:
                  type: object
                  description: Query parameters
                  additionalProperties:
                    type: string
                data:
                  type: string
                  description: Request body data
      responses:
        '200':
          description: Successful API call
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: string
                    description: Raw API response
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
      required:
        - success
        - error

    Customer:
      type: object
      properties:
        Customer:
          type: string
          description: Customer ID
        CustomerName:
          type: string
          description: Customer name
        CustomerFullName:
          type: string
          description: Customer full name

    Product:
      type: object
      properties:
        Product:
          type: string
          description: Product ID
        ProductDescription:
          type: string
          description: Product description
        Language:
          type: string
          description: Language code